Move out pagination into top-lev module
