+++ modules/pagination/client-react/components/PaginationDemoView.native.jsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { View, ScrollView, FlatList, StyleSheet, Text } from 'react-native';
-import { Pagination } from '../../common/components/native';
-
-import translate from '../../../i18n';
+import { Pagination } from '@module/look-client-react-native';
+import { translate } from '@module/i18n-client-react';
 
 class PaginationDemoView extends React.Component {
   render() {
