+++ modules/pagination/client-react/locales/en/translations.json
