+++ modules/pagination/client-react/locales/index.js
