+++ modules/pagination/client-react/locales/ru/translations.json
