+++ modules/upload/client-react/index.native.jsx
@@ -18,7 +18,7 @@ export default new ClientModule({
         screen: createStackNavigator({
           Upload: {
             screen: Upload,
-            navigationOptions: ({ navigation }: any) => ({
+            navigationOptions: ({ navigation }) => ({
               headerTitle: <HeaderTitleWithI18n i18nKey="title" style="subTitle" />,
               headerLeft: (
                 <IconButton iconName="menu" iconSize={32} iconColor="#0275d8" onPress={() => navigation.openDrawer()} />
