+++ packages/mobile/src/index.ts
@@ -5,14 +5,15 @@ import chat from '@module/chat-client-react';
 import contact from '@module/contact-client-react';
 import validation from '@module/validation-common-react';
 import defaultRouter from '@module/router-client-react-native';
-import upload from '@module/upload-client-react';
 
 import post from '../../client/src/modules/post';
 import user from '../../client/src/modules/user';
 import payments from '../../client/src/modules/payments';
-import pagination from '../../client/src/modules/pagination';
 
 import ClientModule from '@module/module-client-react-native';
+
+const upload = require('@module/upload-client-react').default;
+const pagination = require('@module/pagination-client-react').default;
 
 const modules = new ClientModule(
   validation,
